Frontend/COCO-ESPRIT-main/COCO-ESPRIT/requirements.txt
flask
flask-cors
torch
easyocr
pandas
scikit-learn
matplotlib
streamlit
pillow
numpy
nltk
